refactor: move callout components for consistency
