+++ src/components/callout/CalloutCard.vue
@@ -48,8 +48,8 @@ import { format } from 'date-fns';
 import { ref } from '@vue/reactivity';
 import { onBeforeMount } from '@vue/runtime-core';
 import { useI18n } from 'vue-i18n';
-import { formatDistanceLocale } from '../utils/dates/locale-date-formats';
-import { GetCalloutData } from '../utils/api/api.interface';
+import { formatDistanceLocale } from '../../utils/dates/locale-date-formats';
+import { GetCalloutData } from '../../utils/api/api.interface';
 
 const { t } = useI18n();
