+++ src/components/callout/CalloutSummary.vue
@@ -58,12 +58,15 @@
 <script lang="ts" setup>
 import { useI18n } from 'vue-i18n';
 import { computed } from 'vue';
-import { GetCalloutData, GetCalloutDataWith } from '../utils/api/api.interface';
-import AppButton from './forms/AppButton.vue';
-import { formatLocale } from '../utils/dates/locale-date-formats';
-import AppSubHeading from './AppSubHeading.vue';
-import env from './../env';
-import CalloutStatus from './callout/CalloutStatus.vue';
+import {
+  GetCalloutData,
+  GetCalloutDataWith,
+} from '../../utils/api/api.interface';
+import AppButton from '../forms/AppButton.vue';
+import { formatLocale } from '../../utils/dates/locale-date-formats';
+import AppSubHeading from '../AppSubHeading.vue';
+import env from './../../env';
+import CalloutStatus from './CalloutStatus.vue';
 
 const { t } = useI18n();
